requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
psycopg2==2.9.6
datetime==5.2
decorator==5.1.1
pyaes==1.6.1
pyasn1==0.5.0
python-dotenv==1.0.0
rsa==4.9
Telethon==1.28.5
schedule==1.2.0